% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPLBtest.R
\name{BPLBtest}
\alias{BPLBtest}
\title{Función \code{BPLBtest()}}
\usage{
BPLBtest(serie, maxlag, type = c("Ljung", "Box"))
}
\arguments{
\item{serie}{Vector de valores del proceso sobre el que se quiere evaluar incorrelación con tests Ljung-Box o Box-Pierce.}

\item{maxlag}{Entero indicando el máximo orden de rezago temporal hasta el cual se prueba la significancia de la función de autocorrelación.}

\item{type}{Cadena de caracteres para indicar el tipo de prueba, \code{"Ljung"} (por defecto) para las pruebas Ljung-Box y \code{"Box"} para las pruebas Box-Pierce.}
}
\value{
Un objeto \code{data.frame} con tres variables:
\itemize{
\item \code{Xsquared} los valores de los estadísticos de prueba para \verb{m=6, 12,...,[maxlag/6]}.
\item \code{df} los grados de libertad del estadístico de prueba, corresponden a los valores de \verb{m=6, 12,...,[maxlag/6]}.
\item \code{pvalue} los valores P para las pruebas con \verb{m=6, 12,...,[maxlag/6]}.
}
}
\description{
Realiza pruebas Ljung-Box y Box-Pierce para \verb{m=6, 12, 18,...,[maxlag/6]},  con  \code{maxlag} el máximo orden de rezagos. Usa la función \code{Box.test()} de la librería \code{stats}.
}
\examples{
#Simulando serie de un proceso estacionario de media cero con variables correlacionadas
zt <- ts(arima.sim(n=300,list(ar=0.7),sd=4000),freq=1,start=1)
#Aplicando pruebas Ljung-Box con máximo orden de correlación de 36
BPLBtest(serie=zt,maxlag=36)
#Aplicando pruebas Box-Pierce con máximo orden de correlación de 36
BPLBtest(serie=zt,maxlag=36,type="Box")
#Simulando serie de un proceso ruido blanco gaussiano
zt <- ts(rnorm(n=300,sd=4000),freq=1,start=1)
#Aplicando pruebas Ljung-Box con máximo orden de correlación de 36
BPLBtest(serie=zt,maxlag=36)
#Aplicando pruebas Box-Pierce con máximo orden de correlación de 36
BPLBtest(serie=zt,maxlag=36,type="Box")
}
