% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEMENTO.R
\docType{data}
\name{CEMENTO}
\alias{CEMENTO}
\title{Producción trimestral de cemento Portland.}
\format{
\subsection{\code{CEMENTO}}{

Un vector con 155 observaciones.
}
}
\usage{
CEMENTO
}
\description{
Producción trimestral de cemento Portland 1956:Q1 - 1994:Q3,
en miles de toneladas.
}
\keyword{datasets}
