% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatosChippy.R
\docType{data}
\name{DatosChippy}
\alias{DatosChippy}
\title{Datos \code{Chippy} (datos simulados)}
\format{
\subsection{\code{DatosChippy}}{

Un data frame con 210 filas y 2 columnas:
\describe{
\item{Chippy}{Valor de la curva de tendencia en cada tiempo.}
\item{serie}{Valor de la serie en cada tiempo.}
}
}
}
\usage{
DatosChippy
}
\description{
Un conjunto de datos simulados con tendencia \code{T(t)=sin(cos(t)*exp(-t/2))}.
}
\keyword{datasets}
