% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncionDescompLoessv02.R
\name{DescompLoessv02}
\alias{DescompLoessv02}
\title{Función \code{DescompLoessv02()}}
\usage{
DescompLoessv02(
  serie.ajuste,
  tipo.descomp,
  grado = 1,
  criterio,
  h = 1,
  level = 0.95
)
}
\arguments{
\item{serie.ajuste}{Un objeto \code{ts} de serie de tiempo univariada con los valores de la serie a ajustar.}

\item{tipo.descomp}{Tipo de descomposición \code{"additive"} ó \code{"multiplicative"}, en el último caso es bajo modelo parcialmente multiplicativo.}

\item{grado}{Grado del polinomio local a ser ajustado sobre la serie previamente desestacionalizada con la estacionalidad estimada con el filtro de la descomposición clásica de la función R \code{decompose()}, bajo el modelo aditivo o multiplicativo, según valor del argumento \code{tipo.descomp}. Su valor por defecto es 1.}

\item{criterio}{Tipo de criterio de información \code{"aicc"} ó \code{"gcv"}, a usar para la elección del parámetro de suavizamiento en la rutina de loess óptimo mediante la función R \code{loess.as()} de la librería \code{fANCOVA}. Por defecto es \code{"aicc"}.}

\item{h}{Número de predicciones a calcular después del ajuste. Por defecto es 1.}

\item{level}{Nivel de confianza de los intervalos de predicción. Por defecto 0.95.}
}
\value{
Una lista con los siguientes objetos:
\itemize{
\item \code{deltasi} Vector con las estimaciones de los efectos estacionales.
\item \code{alfa.optim} Valor del parámetro de suavizamiento loess óptimo.
\item \code{nep} El número de parámetros equivalentes loess.
\item \code{p} El total de parámetros aproximados en el ajuste de tendencia y estacionalidad.
\item \code{St} Objeto \code{ts} de la estimación de la componente estacional.
\item \code{Tt} Objeto \code{ts} de la estimación de la tendencia loess.
\item \code{ytd} La serie desestacionalizada.
\item \code{fitted} Objeto \code{ts} de la serie estimada combinando aditiva o multiplicativamente, según el tipo de descomposición, las estimaciones de la tendencia y de la estacionalidad.
\item \code{forecast} Objeto \code{ts} multivariado con las predicciones puntuales y por intervalo para los \code{h} períodos pedidos.
\item \code{residuals} Objeto \code{ts} con los residuos de ajuste obtenidos como la diferencia entre los valores observados y los ajustados para la serie.
\item \code{MSE} Estimación de la varianza del error de ajuste.
}
}
\description{
Estima y predice serie estacional aditiva y multiplicativa combinando filtro de la descomposición clásica y la regresión loess.
}
\examples{
#Serie de tiempo con inicio en enero de 1965
ventascompanyX

m <- 6 #tamaño muestra de validación cruzada
n <- length(ventascompanyX)-m #tamaño muestra de ajuste
t <- 1:n #índice de tiempo en periodos de ajuste

#Serie de tiempo en periodos de ajuste
yt <- ts(ventascompanyX[t],frequency=12,start=c(1965,1))

#Ajuste por descomposición multiplicativa y tendencia loess lineal
#se usa criterio AICC para el parámetro de suavizamiento loess
ajusteDLL1 <- DescompLoessv02(serie.ajuste=yt,tipo.descomp="multiplicative",
                              grado=1,criterio="aicc",h=m,level=0.95)

#Gráfico de la serie y su ajuste
plot(ventascompanyX)
lines(fitted(ajusteDLL1),col=2,lwd=2)
legend("topleft",legend=c("Serie","Ajuste"),col=c(1,2),lty=1)

#Pronósticos puntuales y por IPs en los períodos de validación cruzada
ajusteDLL1$forecast

#Estimación de sigma
sigma <- sqrt(ajusteDLL1$MSE)

#residuos del ajuste
diagresTSModel(modelo=ajusteDLL1,RMSE=sigma,all=FALSE,single=FALSE)
}
