% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimarecursiva.R
\name{estimarecursiva}
\alias{estimarecursiva}
\title{Funcion \code{estimarecursiva()}}
\usage{
estimarecursiva(
  respuesta,
  dataX,
  min.n,
  nrow1 = 2,
  ncol1 = 1,
  nrow2 = 2,
  ncol2 = 2,
  plot.recursive = TRUE
)
}
\arguments{
\item{respuesta}{Un vector numérico o serie de tiempo con los valores de la respuesta del modelo de regresión lineal múltiple.}

\item{dataX}{Un \code{data.frame} cuyas columnas son los predictores del modelo de regresión lineal múltiple, el número de filas debe ser igual a la longitud de la variable \code{respuesta}.}

\item{min.n}{Un escalar para indicar el mínimo tamaño de muestra con el que debe iniciar la estimación recursiva, el cual no puede ser inferior a \code{ncol(dataX) + 2}.}

\item{nrow1, ncol1}{Valores enteros que definen el número de \verb{filas, columnas} en el diseño de la ventana gráfica para representar las gráficas de los residuales recursivos y del estadístico \code{CUSUMt-Recursivo}. Si ambos valores son iguales a 1, cada gráfica sale en una ventana por separado.}

\item{nrow2, ncol2}{Valores enteros que definen el número de \verb{filas, columnas} en el diseño de la ventana gráfica para representar las gráficas de las estimaciones recursivas de los parámetros cuando el argumento \code{plot.recursive = TRUE}. Cuando el número de gráficas a ser representadas  supera al número de celdas definido en el diseño de la ventana de salida, entonces se generarán las ventanas gráficas adicionales que sean necesarias con el mismo diseño.}

\item{plot.recursive}{Argumento lógico, sus valores posibles son \code{TRUE} (por defecto) y \code{FALSE}, para indicar si se desean o no las gráficas de las estimaciones recursivas de los parámetros del modelo de regresión lineal múltiple.}
}
\value{
La función muestra en la consola \code{R} la estimación del modelo global sobre el total de observaciones leídas en el vector \code{respuesta} y los resultados del test \code{CUSUMt-recursivo}. Genera también el gráfico de los residuos recursivos y el gráfico del estadístico \code{CUSUMt-recursivo}. Además, produce un objeto tipo lista con los siguientes componentes:
\itemize{
\item \code{n} Una matriz de una sola columna con sus filas siendo los valores de tamaño de muestra desde \code{min.n} hasta \code{length(respuesta)}.
\item \code{estimacion_recursiva} Un arreglo de matrices, donde cada matriz tiene tres columnas: la estimación de los parámetros y los límites inferior y superior de confianza del 95\% de cada parámetro. El número de matrices es igual al total de tamaños de muestras entre \code{min.n} y \code{length(respuesta)}.
\item \code{ajusteglobal} Un objeto tipo \code{lm()} del ajuste de la \code{respuesta} vs. los predictores en \code{dataX}.
\item \code{resid_recursivos} Un vector de residuos recursivos.
\item \code{test_CUSUM} Los resultados (estadístico de prueba y valor-P) del test \code{CUSUMt-recursivo}.
}
}
\description{
Realiza la estimación recursiva de un modelo de regresión lineal, obteniendo además los residuos recursivos, la gráfica \code{CUSUMt-Recursivo} y el test \code{CUSUMt-Recursivo} mediante las funciones \code{recresid()}, \code{efp()} y \code{sctest()}, respectivamente, de la librería \code{strucchange}.
}
\examples{
library(forecast)
datos <- ts(CEMENTO,frequency=4,start=c(1956,1)) # serie CEMENTO con todos los datos
#Variables para modelo log cuadrático estacional con indicadoras, nivel referencia trimestre 4
t <- 1:length(datos) #Definiendo índice de tiempo para todos los datos
poli <- Mipoly(tiempo=t,grado=2)
trimestre <- seasonaldummy(datos) #Creando variables indicadoras, nivel de  referencia trimestre 4
X <- data.frame(poli,trimestre) #Matriz de predictores
resultados <- estimarecursiva(respuesta=log(datos),dataX=X,min.n=24,
              nrow1=2,ncol1=1,nrow2=2,ncol2=2,plot.recursive=TRUE)
}
