% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filtrolineal.R
\name{Filtrolineal}
\alias{Filtrolineal}
\title{Función \code{Filtrolineal()}}
\usage{
Filtrolineal(serie, pesos.wi, tipo, h = 1, level = 0.95)
}
\arguments{
\item{serie}{Un objeto tipo \code{ts} con los valores de la serie a ajustar y pronosticar.}

\item{pesos.wi}{Vector de pesos de las medias moviles para la estimacion del nivel. Para medias moviles simples unilaterales debe ser de longitud \code{m+1}, para \verb{i=0,1,...,m}, y para bilaterales debe ser de longitud \verb{2m+1}, para \verb{i=-m,...,m}.}

\item{tipo}{El tipo de media móviles para la estimacion del nivel, \code{tipo=1} para medias móviles simples unilaterales, \code{tipo=2} para medias moviles simples bilaterales.}

\item{h}{Número de periodos futuros a pronosticar, por defecto es 1.}

\item{level}{El nivel de confianza para los intervalos de pronóstico, por defecto es 0.95.}
}
\value{
Una lista con los siguientes componentes:
\itemize{
\item \code{nivel} Objeto tipo \code{ts} de las medias moviles calculadas con la estrategia circular.
\item \code{fitted} Objeto tipo \code{ts} con los valores ajustados de la serie.
\item \code{residuals} Objeto tipo \code{ts} con los valores de la serie de tiempo de los residuos del ajuste.
\item \code{forecast} Objeto tipo \code{mts} con los valores de los pronósticos puntuales y por Intervalos para los \code{h} periodos.
\item \code{RMSE} Estimacion de la raíz cuadrada del MSE del ajuste.
}
}
\description{
Ajusta y pronostica series no estacionales a través de filtros lineales basados en medias moviles simples unilaterales o bilaterales, usando la estrategia circular.
}
\examples{
#Serie de tasa de variación trimestral del PIB in UK 1990:Q1 a 2004:Q1
datos <- ts(GDPchange,freq=4,start=c(1990,1))
#grafico de la serie
plot(datos,type="b",pch=19,lwd=3)

m <- 3 #tamano muestra de validacion cruzada
n <- length(datos)-m #tamano muestra de ajuste
t <- 1:n
yt <- ts(datos[t],freq=4,start=c(1990,1))
tnuevo <- (n+1):length(datos)
ytnuevo <- ts(datos[tnuevo],freq=4,start=c(2003,3)) #pronosticos inician en Q3-2003

#Ajuste y pronósticos con Filtro de Henderson de 2m+1=7 términos
w7 <- c(-0.059,0.059,0.294,0.413,0.294,0.059,-0.059) #Vector de pesos wi
FH7 <- Filtrolineal(serie=yt,pesos.wi=w7,tipo=2,h=m) #Obtención de ajustes y pronosticos
ythatFH7 <- fitted(FH7);ythatFH7 #Valores ajustados

#Serie y su ajuste
plot(datos,lwd=2)
lines(ythatFH7,col=2,lwd=2)
legend("bottomright",legend=c("Obs.","FH7"),col=c(1,2),lty=1,cex=2,lwd=2)
#RMSE de ajuste
RMSEFH7 <- FH7$RMSE
RMSEFH7

#pronósticos puntuales y por IPs del 95\%
pronFH7 <- FH7$forecast
pronFH7
#Precisión pronósticos puntuales
forecast::accuracy(FH7$forecast[,1],ytnuevo)
#Precisión pronósticos por IPs
precisintervals(real=ytnuevo,LIP=FH7$forecast[,2],LSP=FH7$forecast[,3])
#Gráfico serie de residuos de ajuste
ymin <- min(residuals(FH7),-2*RMSEFH7,2*RMSEFH7,na.rm=TRUE)
ymax <- max(residuals(FH7),-2*RMSEFH7,2*RMSEFH7,na.rm=TRUE)
plot(residuals(FH7),ylim=c(ymin,ymax))
abline(h=c(-2*RMSEFH7,0,2*RMSEFH7),col=2)
legend("bottomright",legend="FH7",cex=2)
}
