% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDPchange.R
\docType{data}
\name{GDPchange}
\alias{GDPchange}
\title{Variación porcentual del PIB de UK.}
\format{
\subsection{\code{GDPchange}}{

Un vector con 57 observaciones.
}
}
\usage{
GDPchange
}
\description{
Variación porcentual trimestral del PIB en el Reino Unido, 1990:1Q - 2004:1Q
}
\keyword{datasets}
