% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpdeltas.R
\name{interpdeltas}
\alias{interpdeltas}
\title{\verb{Función interpdeltas()}}
\usage{
interpdeltas(
  modelo,
  ordenp = 0L,
  ordenq = 0L,
  ordenP = 0L,
  ordenQ = 0L,
  gradopoly,
  aditivo = TRUE,
  plotit = TRUE
)
}
\arguments{
\item{modelo}{El nombre del objeto R con el modelo ajustado. Objetos tipo \code{lm()}, \code{nls()} o \code{Arima()}.}

\item{ordenp, ordenq, ordenP, ordenQ}{Argumentos de valor entero, para especificar los valores de los ordenes \code{p}, \code{q}, \code{P}, \code{Q} de modelos \code{ARMA(p,q)(P,Q)[s]}. Por defecto están fijos en 0, es decir, asumiendo error RB. Si el error es un \code{AR(p)} el usuario debe especificar el valor de \code{p}; si el error es \code{MA(q)} debe especificar el valor de \code{q}; si es \code{ARMA(p,q)} debe especificar \code{p} y \code{q}, etc.}

\item{gradopoly}{El grado del polinomio global. Siempre debe ser especificado.}

\item{aditivo}{Argumento lógico \verb{(TRUE, FALSE)}, por defecto es \code{TRUE}, este valor debe usarse en modelos polinomiales estacionales con indicadoras y error RB o error ARMA. En modelos log polinomiales estacionales con error RB o error ARMA, o modelos exponenciales polinomiales estacionales con error RB, este argumento debe especificarse igual a \code{FALSE}.}

\item{plotit}{Argumento lógico \verb{(TRUE, FALSE)}, por defecto es \code{TRUE} indicando que se desea la gráfica de los efectos estacionales para su interpretación en la escala de la serie, según si el modelo es aditivo o multiplicativo.}
}
\value{
Con \code{plotit=TRUE}, despliega la gráfica correspondiente y en la consola R el vector de parámetros estacionales estimados (modelos aditivos) o de los valores exponenciados de estas estimaciones y multiplicados por 100\% (modelos multiplicativos). Además, crea un objeto lista con los siguientes componentes:
\itemize{
\item \code{periodo} Un vector de valores enteros de 1 a \code{s} (la longitud del año calendario).
\item \code{deltasi} Un vector con las estimaciones de los parámetros asociados a las indicadoras y adiciona el valor de cero al final. Este vector aparece cuando \code{aditivo=TRUE}.
\item \code{expdeltasi100} Un vector con los valores exponenciados de las estimaciones de los parámetros asociados a las indicadoras, multiplicados por 100 y adiciona el valor de 100 al final. Este vector aparece cuando \code{aditivo=FALSE}.
}
}
\description{
Extrae las estimaciones y construye gráfico de los efectos estacionales en modelos de regresión global estacionales que usan indicadoras con nivel de referencia el \code{s} ésimo período calendario. Particularmente en modelos polinomiales estacionales con error RB y con error ARMA, modelos log polinomiales estacionales con error RB y con error ARMA y modelos exponenciales polinomiales estacionales con error RB. En los modelos polinomiales estacionales, bien sea con error RB o ARMA, extrae los valores estimados de los parámetros asociados a las variables indicadoras y crea la gráfica de estos valores vs. período calendario, asignando valor de cero en el último período. En los modelos log polinomiales estacionales con error RB o ARMA y en los modelos exponenciales polinomiales estacionales con error RB, extrae los valores estimados de los parámetros asociados a las variables indicadoras, pero es necesario informar que estos modelos son multiplicativos para que sean exponenciadas estas estimaciones, las multiplique por 100\% y las grafique contra el período calendario, asignando el valor de 100\% al último periodo del año.
}
\examples{
library(forecast)
#Ej. 1 Modelo polinomial estacional con indicadoras (nivel ref=12)
datos1 <- salario
n1 <- length(datos1)-12
t1 <- 1:n1
yt1 <- ts(datos1[t1],freq=12,start=start(datos1))
poli1 <- Mipoly(tiempo=t1,grado=4)
mes <- seasonaldummy(yt1)
X1 <- data.frame(poli1,mes)
modelo1 <- lm(yt1~.,data=X1)
summary(modelo1)
#Extrayendo y graficando las estimaciones de los coeficientes de regresión delta_i
interpdeltas(modelo1,gradopoly=4,aditivo=TRUE,plotit=TRUE)

#Ej. 2 log cúbico estacional con indicadoras (nivel ref=Q4), error ARMA(1,2)x(0,1)[4]
datos2 <- ts(CEMENTO,fre=4,start=c(1956,1))
n2 <- length(datos2)-4
t2 <- 1:n2
yt2 <- ts(datos2[t2],freq=4,start=start(datos2))
poli2 <- Mipoly(tiempo=t2,grado=3)
trimestre <- seasonaldummy(yt2)
X2 <- data.frame(poli2,trimestre)
modelo2 <- Arima(log(yt2),order=c(1,0,2),
seasonal=list(order=c(0,0,1)),xreg=as.matrix(X2),method="ML")
k2 <- length(coef(modelo2)[coef(modelo2)!=0]);k2 #número de parámetros
df2 <- n2-k2 #grados de libertad
lmtest::coeftest(modelo2,df=df2)
interpdeltas(modelo=modelo2,ordenp=1,ordenq=2,ordenP=0,ordenQ=1,
             gradopoly=3,aditivo=FALSE,plotit=TRUE)

#Ej. 3, exponencial cúbico estacional con indicadoras (nivel ref=Q4)
modelo3 <- regexponencialv02(respuesta=yt2,data=X2)
summary(modelo3)
interpdeltas(modelo=modelo3,gradopoly=3,aditivo=FALSE,plotit=TRUE)
}
