% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoessOptimo.R
\name{LoessOptimo}
\alias{LoessOptimo}
\title{Función \code{LoessOptimo()}}
\usage{
LoessOptimo(serie, grado = 1, criterio, h = 1, level = 0.95)
}
\arguments{
\item{serie}{Un objeto tipo \code{ts} de serie de tiempo univariada con los valores de la serie a ajustar.}

\item{grado}{Grado del polinomio local a ser ajustado sobre la serie. \code{grado=1} es para loess lineal y \code{grado=2} es para loess cuadrático. Su valor por defecto es 1.}

\item{criterio}{Tipo de criterio de información \code{"aicc"} ó \code{"gcv"}, a usar para la elección del parámetro de suavizamiento en la rutina de loess óptimo mediante la función R \code{loess.as()} de la librería \code{fANCOVA}. Por defecto es \code{"aicc"}.}

\item{h}{Número de predicciones a calcular después del ajuste. Por defecto es 1.}

\item{level}{Nivel de confianza de los intervalos de predicción. Por defecto 0.95.}
}
\value{
Una lista con los siguientes objetos:
\itemize{
\item \code{alfa.optim} valor del parámetro de suavizamiento loess óptimo.
\item \code{nep} El número de parámetros equivalentes loess.
\item \code{MSE} Estimación de la varianza del error de ajuste.
\item \code{fitted} Objeto \code{ts} de la serie estimada.
\item \code{residuals} Objeto \code{ts} con los residuos de ajuste obtenidos como la diferencia entre los valores observados y los ajustados para la serie.
\item \code{forecast} Objeto \code{ts} multivariado con las predicciones puntuales y por intervalo para los \code{h} períodos pedidos.
}
}
\description{
Para ajuste y pronóstico de una serie no estacional mediante loess óptimo según criterios de información.
}
\examples{
serie <- ts(DatosChippy$serie,freq=12,start=c(1990,1))
curva <- ts(DatosChippy$Chippy,freq=12,start=c(1990,1))

#Gráfica de la serie y tendencia verdadera
plot(serie,lwd=1)
lines(curva,col=2,lwd=2)
legend("topright",legend=c(expression(sin(cos(t)*exp(-t/2)))),col=2,lty=1,lwd=2)

#Ajuste con validación cruzada con últimos 4 datos
h <- 4
n <- length(serie)-h
t <- 1:n #tiempos de ajuste
#valores a ajustar
yt <- ts(serie[t],freq=frequency(serie),start=start(serie))
tnuevo <- (n+1):length(serie) #tiempos de predicciones
#Fecha inicio predicciones
startpred=end(ts(serie[1:(n+1)],freq=frequency(serie),start=start(serie)))
#Valores observados en tiempos de predicciones
ytnuevo <- ts(serie[tnuevo],freq=frequency(serie),start=startpred)

ajusteLoess <- LoessOptimo(serie=yt,grado=2,criterio="aicc",h=h,level=0.95)
ythat <- fitted(ajusteLoess) #serie ajustada

#Serie y su ajuste
plot(serie)
lines(curva,col=2,lwd=2) #Tendencia verdadera
lines(ythat,col="blue1",lwd=2)
legend("topright",legend=c("Tendencia verdadera","Ajuste"),
col=c(2,"blue1"),lty=1,lwd=2)

#Número aproximado de parámetros
p <- round(ajusteLoess$nep)
p
#AIC y BIC aproximados version exp(C*n(p))
Criterios <- critinfresid(residuales=residuals(ajusteLoess),npar=p)
Criterios

#Pronósticos con IPs del 95\% de confianza
pronosticos <- ajusteLoess$forecast
pronosticos

#MSE aproximado del ajuste
MSE <- ajusteLoess$MSE

#Gráficos de residuos de ajuste
diagresTSModel(modelo=ajusteLoess,RMSE=sqrt(MSE),all=FALSE,single=FALSE)
}
