% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mipoly.R
\name{Mipoly}
\alias{Mipoly}
\title{Función \code{Mipoly()}}
\usage{
Mipoly(tiempo, grado)
}
\arguments{
\item{tiempo}{Un vector con los valores del indice de tiempo t.}

\item{grado}{El grado p del polinomio deseado, debe ser un valor entero positivo a partir de 1.}
}
\value{
Un \code{data.frame} cuyas variables son \code{t} y sus potencias denominadas \code{t2}, \code{t3}, \code{...}, \code{tp}, según el grado polinomial p.
}
\description{
Genera las variables de los polinomios para modelos de regresión con funciones de tendencia polinomiales. Se basa en la función \code{poly()} de \code{stats}.
}
\examples{
t <- 1:20 #índice de tiempo primeros 20 valores
poli2 <- Mipoly(tiempo=t,grado=2) #para predictores con polinomio de grado 2
poli2
poli3 <- Mipoly(tiempo=t,grado=3) #para predictores con polinomio de grado 3
poli3
}
