% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mytrigon.R
\name{Mytrigon}
\alias{Mytrigon}
\title{Función \code{Mytrigon()}}
\usage{
Mytrigon(tiempo, Frecuencias, indicej)
}
\arguments{
\item{tiempo}{Vector de valores enteros positivos con el índice de tiempo para los cuales se debe calcular las funciones trigonométricas.}

\item{Frecuencias}{Vector con las frecuencias \code{Fj} correspondientes a las ondas sinusoidales armónicas a considerar.}

\item{indicej}{Vector de valores enteros indicando el sub índice \code{j} de las funciones trigonométricas para cada frecuencia en \code{Frecuencias} de manera que a \code{Frecuencias[i]} le asigna el indice \code{indicej[i]}.}
}
\value{
Un \code{data.frame} con las variables \code{sinFj}, \code{cosFj}, siendo \code{j} el valor asignado en \code{indicej}. Para la frecuencia 1/2 solo crea la componente coseno, que corresponde a \code{cos(pi*tiempo)}. Los pares de variables \code{sinj}, \code{cosj} aparecen en orden de menor a mayor frecuencia pero conservando el sufijo \code{j} asignado en \code{indicej}.
}
\description{
Genera las variables trigonométricas \code{sin(2*pi*Fj*tiempo)} y \code{cos(2*pi*Fj*tiempo)} en frecuencias Fj.
}
\examples{
t <- 1:20 #índice de tiempo primeros 20 valores
#Trigonométricas en Fj=j/12 para j=1 a 6 y F7=0.35
trigono <- Mytrigon(tiempo=t,Frecuencias=c(c(1:6)/12,0.35),indicej=c(1,2,3,4,5,6,7))
trigono #Variables sinF7, cosF7 corresponden a la frecuencia F7=0.35
}
