% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precisintervals.R
\name{precisintervals}
\alias{precisintervals}
\title{Función \code{precisintervals()}}
\usage{
precisintervals(real, LIP, LSP, alpha = 0.05)
}
\arguments{
\item{real}{Vector numérico con los valores observados en los periodos de pronóstico.}

\item{LIP}{Vector numérico con los valores de los límites inferiores de los IPs.}

\item{LSP}{Vector numérico con los valores de los límites superiores de los IPs.}

\item{alpha}{Escalar entre 0 y 1 correspondiendo al nivel \code{1-alpha} de confianza.}
}
\value{
Un vector con \code{AmplitudProm} la amplitud promedio, \verb{Cobertura(\%)} la cobertura (porcentaje) y \code{ScoreProm} el Score promedio de los IPs.
}
\description{
Evalúa precisión de intervalos de predicción (IPs).
}
\examples{
Real <- c(181958.0,185303.0,183429.0)
Pronost <- data.frame(rbind(c(175546.9,163886.4,188037.2),
c(175938.6,164246.2,188463.5),c(176323.1,164599.1,188882.2)))
names(Pronost) <- c("Pred","LimInf","LimSup")
Pronost
precisintervals(real=Real,LIP=Pronost[,2],LSP=Pronost[,3])
}
