% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictexpo.R
\name{predictexpo}
\alias{predictexpo}
\title{Función \code{predictexpo()}}
\usage{
predictexpo(
  modelo,
  new.data,
  level = 0.95,
  interval = c("none", "confidence", "prediction")
)
}
\arguments{
\item{modelo}{Objeto tipo \code{nls()} generado por la función \code{regexponencialv02()}.}

\item{new.data}{\code{data.frame} con los valores de los predictores en los tiempos de predicción. El nombre de las variables debe ser igual al usado en el ajuste del modelo.}

\item{level}{Un valor en (0,1) para el nivel de confianza de los intervalos. Por defecto es 0.95.}

\item{interval}{Una cadena de caracteres correspondiendo a uno de los siguientes casos: \code{"none"}, \code{"confidence"}, \code{"prediction"}. Por defecto es \code{"none"}, es decir, calcula solo estimaciones o pronósticos puntuales. Para intervalos de confianza en la respuesta media use \code{"confidence"} y para intervalos de predicción use \code{"prediction"}.}
}
\value{
Un objeto tipo \code{data.frame} con tres variables:
\itemize{
\item \code{forecast} La estimación o pronóstico puntual de la respuesta.
\item \code{lower}, \code{upper} Los límites inferior y superior de los intervalos pedidos.
}
}
\description{
Para la predicción puntual y por IPs sobre modelos de regresión exponencial ajustados con la función \code{regexponencialv02()}. También calcula I.Cs para la respuesta media en lugar de IPs.
}
\examples{
#Simulando serie aditiva con modelo de tendencia exponencial lineal y errores correlacionados
n <- 250
t <- 1:n
error <- arima.sim(n=n,list(ar=0.8),sd=108)
Yt <- ts(exp(6.5+0.0105*t)+error,freq=1,start=1)
plot.ts(Yt)
lines(t,exp(6.5+0.0105*t),type="l",col="red2")

#Matriz de predictores
X <- Mipoly(tiempo=t,grado=1)
#Ajuste del modelo exponencial lineal
modelo <- regexponencialv02(respuesta=Yt,data=X)
summary(modelo)
#cinco tiempos de predicción
Xnuevo=Mipoly(tiempo=(n+1):(n+5),grado=1)
#Predicciones puntuales y por IPs del 95\% de confianza
predictexpo(modelo=modelo,new.data=Xnuevo,interval="prediction")
}
