% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruebaDW1.R
\name{pruebaDW1}
\alias{pruebaDW1}
\title{Función \code{pruebaDW1()}}
\usage{
pruebaDW1(modelo)
}
\arguments{
\item{modelo}{Un objeto tipo \code{lm} en el que se guardó el ajuste de un MRLM.}
}
\value{
un objeto \code{data.frame} con las siguientes variables:
\itemize{
\item \verb{rho(1) estimado} estimación de la autocorrelación de orden 1 \code{rho(1)} usando los residuales del modelo.
\item \verb{Estadistico D-W} El valor del estadístico de la prueba.
\item \verb{VP H1: rho(1)>0} El valor P para el test \code{H0: rho(1)=0}, vs. \code{H1:rho(1)>0}.
\item \verb{VP H1: rho(1)<0} El valor P para el test \code{H0: rho(1)=0}, vs. \code{H1:rho(1)<0}.
}
}
\description{
Realiza test Durbin-Watson para autocorrelación de orden 1 en un MRLM. Usa la función \code{durbinWatsonTest()} de la librería \code{car}.
}
\examples{
#Simulando serie con modelo de tendencia lineal y errores RB normales
t <- 1:300
yt <- 3500+200*t+rnorm(300,sd=4000)
yt <- ts(yt,freq=12,start=c(1980,1))
#Gráfico de la serie simulada
plot(yt)
#Ajuste del modelo de tendencia lineal y errores RB normales
mrl <- lm(yt~t)
summary(mrl)
#Aplicando función para test Durbin-Watson
pruebaDW1(modelo=mrl)
#Simulando serie con modelo de tendencia lineal y errores correlacionados
zt <- 3500+200*t+arima.sim(n=300,list(ar=0.7),sd=4000)
zt <- ts(zt,freq=12,start=c(1980,1))
#Gráfico de la serie simulada
plot(zt)
#Ajuste del modelo de tendencia lineal y errores RB normales
mrl2 <- lm(zt~t)
summary(mrl2)
#Aplicando función para test Durbin-Watson
pruebaDW1(modelo=mrl2)
}
