% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regexponencialv02.R
\name{regexponencialv02}
\alias{regexponencialv02}
\title{Función \code{regexponencialv02()}}
\usage{
regexponencialv02(respuesta, data, control = stats::nls.control())
}
\arguments{
\item{respuesta}{El vector de valores de la variable respuesta.}

\item{data}{Un objeto \code{data.frame} con los predictores a usar.}

\item{control}{Una lista opcional de configuraciones de control. Consulte \code{nls.control()} para los valores de control configurables y su efecto.}
}
\value{
Son los mismos predeterminados para los objetos tipo \code{nls} generados por la función R \code{nls()}.
}
\description{
Ajusta modelos de regresión de la forma \code{g(x)=exp(t(x)\%*\%beta)}, con \code{t(x)} la transpuesta de un vector de predictores \code{x} y \code{beta} el vector de parámetros.
}
\examples{
#Simulando serie aditiva con modelo de tendencia exponencial lineal y errores iid normales
n <- 250
t <- 1:n
error <- arima.sim(n=n,list(ar=0.8),sd=108)
Yt <- ts(exp(6.5+0.0105*t)+error,freq=1,start=1)
plot.ts(Yt)
lines(t,exp(6.5+0.0105*t),type="l",col="red2")

#Matriz de predictores
X <- Mipoly(tiempo=t,grado=1)
#Ajuste del modelo exponencial lineal
modelo <- regexponencialv02(respuesta=Yt,data=X)
summary(modelo)

#Gráficos de residuos de ajuste
diagresTSModel(modelo=modelo,RMSE=summary(modelo)$sigma,all=FALSE,single=FALSE)
}
