% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSALES.R
\docType{data}
\name{RSALES}
\alias{RSALES}
\title{Ventas mensuales al por menor}
\format{
\subsection{\code{RSALES}}{

Un data frame con 493 filas y 1 columna:
\describe{
\item{rsales}{Ventas al por menor}
}
}
}
\source{
Diebold, F. (2001), Elementos de Pronósticos.
}
\usage{
RSALES
}
\description{
Ventas mensuales del comercio por menor en Estados Unidos, 1954:1-1995:1,
en dólares nominales.
}
\keyword{datasets}
