% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salario.R
\docType{data}
\name{salario}
\alias{salario}
\title{Índice de salario real sector manufactura sin trilla de café, año base 2001.}
\format{
\subsection{\code{salario}}{

Un objeto \code{ts} univariado con 257 observaciones.
}
}
\source{
DANE (MMM).
}
\usage{
salario
}
\description{
Valores mensuales del índice de salario real en Colombia sector de manufactura sin trilla de café, 1990:1 a 2011:5,
en puntos del índice
}
\keyword{datasets}
