% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectModelv02.R
\name{SelectModelv02}
\alias{SelectModelv02}
\title{Función \code{SelectModelv02()}}
\usage{
SelectModelv02(
  z,
  lag.max = 15,
  Criterion = "default",
  Best = 3,
  Candidates = 5,
  t = "default"
)
}
\arguments{
\item{z}{Una serie de tiempo.}

\item{lag.max}{Un entero indicando el máximo orden autorregresivo para el proceso \code{AR(p)}.}

\item{Criterion}{El criterio de información a usar para la selección del orden \code{p}. Por defecto es \code{"BIC"} (Bayesian Information Criterion). Otras opciones son \code{"AIC"} (Akaike Information Criterion), \code{"UBIC"} (Uncertainty-Penalized Bayesian Information Criterion), \code{"EBIC"} (Extended BIC), y \code{"GIC"} (Generalizad Information Criterion).}

\item{Best}{El número final de modelos a ser seleccionados.}

\item{Candidates}{El número de inicialmente seleccionados usando el criterio aproximado.}

\item{t}{Parámetro de ajuste en los criterios\code{"EBIC"}, \code{"GIC"}.}
}
\value{
La función crean un objeto tipo \code{"matrix"}, \code{"array"} de \code{Best} filas por 3 columnas, con los siguientes valores en sus columnas: \code{p} que corresponde a los valores del orden autorregresivo para los \code{Best} primeros casos de mínimo valor del criterio usado; los valores del criterio de información correspondiente, el exacto: \code{Criterion-Exact} y el aproximado: \code{Criterion-Approx}, donde \code{Criterion} corresponde al acrónimo del criterio de información usado.
}
\description{
Usando criterios de información selecciona el mejor modelo AR(p) completo. Está basado en la Función SelectModel() de la librería FitAR (ya no disponible en Cran-R) de A.I. McLeod and Y. Zhang.
}
\examples{
#Simulando serie de un proceso AR(1) estacionario de media cero
zt <- ts(arima.sim(n=300,list(ar=0.7),sd=4000),freq=1,start=1)
SelectModelv02(z=zt,Criterion="AIC")
SelectModelv02(z=zt,Criterion="BIC")

#Simulando serie de un proceso AR(6) estacionario de media cero
xt <- arima.sim(n=300,list(ar=c(0.65,0.32,-0.27,0.25,0,-0.23)),sd=0.025)
SelectModelv02(z=xt,Criterion="AIC")
SelectModelv02(z=xt,Criterion="BIC")
SelectModelv02(z=xt,Criterion="EBIC")
SelectModelv02(z=xt,Criterion="UBIC")
SelectModelv02(z=xt,Criterion="GIC")
}
