% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ventascompanyX.R
\docType{data}
\name{ventascompanyX}
\alias{ventascompanyX}
\title{Ventas mensuales de la compañía X.}
\format{
\subsection{\code{ventascompanyX}}{

Un objeto \code{ts} univariado con 77 observaciones.
}
}
\source{
O.D. Anderson, 1976 and O’Donovan, 1983.
}
\usage{
ventascompanyX
}
\description{
Serie ficticia de las Ventas mensuales de una compañía X, 1965:1 a 1971:5,
en cientos de dólares.
}
\keyword{datasets}
